+++ .github/workflows/travis.yml
@@ -2,9 +2,9 @@ name: rucron
 
 on:
   push:
-    branches: [ master ]
+    branches: [ main ]
   pull_request:
-    branches: [ master ]
+    branches: [ main ]
 
 env:
   CARGO_TERM_COLOR: always
@@ -43,8 +43,8 @@ jobs:
         override: true
         profile: minimal
     - uses: Swatinem/rust-cache@v1
-    - name: cargo doc
-      run: cargo doc --all-features --no-deps
+    - name: Default doc
+      run: cargo doc --no-deps
 
   test:
     name: Test
@@ -54,5 +54,5 @@ jobs:
     - uses: actions/checkout@v2
     - name: Build
       run: cargo build --verbose
-    - name: Run tests
+    - name: Default tests
       run: cargo test --verbose -- --test-threads 1
