+++ Cargo.toml
@@ -8,8 +8,17 @@ description = "A Rust Job Scheduling Crate."
 license = "MIT"
 repository = "https://github.com/goldwind-ting/rucron/"
 
+[features]
+default = ["tokio"]
+tokio = ["dep:tokio"]
+smol = ["dep:async-channel", "dep:async-lock", "dep:smol"]
+
+
 [dependencies]
-tokio = { version = "1.23.0", features = ["time", "full"] }
+tokio = { version = "1.23.0", features = ["time", "full"], optional = true }
+async-channel = { version = "1.8.0",  optional = true }
+async-lock = { version = "2.6.0",  optional = true }
+smol = { version = "1.3.0",  optional = true }
 futures = "0.3.25"
 lazy_static = "1.4.0"
 signal-hook = "0.3.14"
@@ -22,10 +31,13 @@ thiserror = "1.0.38"
 serde_json = "1.0.91"
 serde = { version = "1.0.151", features = ["derive"] }
 rayon = "1.6.1"
+select-macro = "0.2.0"
+
 
 [dev-dependencies]
 redis = "0.22.1"
 fastrand = "1.8.0"
+tokio = { version = "1.23.0", features = ["time", "full"] }
 
 [[example]]
 name = "redlocker"
